Address PR review comments (third pass): return `-seconds` suffix

+++ pkg/charts/linkerd2/values.go
@@ -95,7 +95,7 @@ type (
 		Resources              *Resources    `json:"resources"`
 		Trace                  *Trace        `json:"trace"`
 		UID                    int64         `json:"uid"`
-		WaitBeforeExitSeconds  int32         `json:"waitBeforeExitSeconds"`
+		WaitBeforeExitSeconds  uint64        `json:"waitBeforeExitSeconds"`
 	}
 
 	// ProxyInit contains the fields to set the proxy-init container
